Add files via upload
The main goal of this work is to detect the internal contours of the left ventricle (LV) on a Magnetic Resonance Image (MRI) using the Snakes segmentation method. For that, we started initializing the weighting parameters. After that, we used a gaussian filtering to inhance the images contrast then gradient forces to extract gradient vectors. Finally, next step consists in detecting the internal contour.